+++ modules/utility/utils.mjs
@@ -1,5 +1,7 @@
-export default  class Utils {
- static findActor(actorId) {
+import QuestFolder from "../entities/quest-folder.mjs";
+
+export default class Utils {
+  static findActor(actorId) {
     let actor = game.actors.get(actorId);
     if (actor === undefined || actor === null) {
       actor = game.actors.find(a => a.name === actorId);
@@ -28,17 +30,53 @@ export default  class Utils {
   }
 
   /**
-   * Update Macros to use new global API
+   * Update Quests to use newer, flexible data format
    */
-  static updateMacros() {
-    let macros = game.macros.filter(e => e.visible);
+  static updateQuests() {
+    const rootFolder = QuestFolder.get('root');
+    let questDirs = {
+      active: '_fql_active',
+      completed: '_fql_completed',
+      failed: '_fql_failed',
+      hidden: '_fql_hidden'
+    };
 
-    macros.forEach((macro) => {
-      let data = duplicate(macro);
-      data.command = data.command.replace(/game\.questlog\./g, 'QuestLog.');
-      data.command = data.command.replace(/game\.quests\./g, 'Quests.');
+    for (let key in questDirs) {
+      const value = questDirs[key];
+      let folder = game.journal.directory.folders.find(f => f.name === value);
 
-      macro.update(data);
-    });
+      folder.content.forEach(entry => {
+        let content = entry.data.content;
+        content = JSON.parse(content);
+
+        let isAvailable = false;
+        if (key === 'hidden') {
+          isAvailable = (entry.data.permission.default === 2 || content.personal);
+        }
+        content.id = entry._id;
+        content.status = key;
+        content.giver = null;
+
+        if (isAvailable) {
+          content.status = 'available';
+        }
+
+        let actor = content.actor || null;
+        if (actor !== null) {
+          let actorE = game.actors.get(actor);
+          if (actorE) {
+            content.giver = actorE.uuid;
+          }
+        }
+        delete content.actor;
+
+        content = JSON.stringify(content);
+        entry.update({content: content, folder: rootFolder}, {diff: false});
+        console.log(entry);
+        console.log(JSON.parse(content));
+      });
+
+      folder.delete();
+    }
   }
 };