+++ changelog.md
@@ -7,6 +7,7 @@
 * Added Configuration setting to decide if hidden tasks should be counted
 * Allowed to drag&drop Quests from Quest Log to create rich-text entity links
 * Added "Copy Link" button on Quest Preview window
+* Migrated from quest folders to quest statuses. 
 
 ### v0.3.3
 * Fixed editable inputs cutting off text after double quotation character